+++ docs/changelog.rst
@@ -7,11 +7,16 @@ Version Numbering
 TinyDB follows the SemVer versioning guidelines. For more information,
 see `semver.org <http://semver.org/>`_
 
+**v1.4.0** (2014-07-XX)
+^^^^^^^^^^^^^^^^^^^^^^^
+
+- Added ``insert_multiple`` function (see `issue #8 <https://github.com/msiemens/tinydb/issues/8>`_).
+
 **v1.3.0** (2014-07-02)
 ^^^^^^^^^^^^^^^^^^^^^^^
 
 - Fixed `bug #7 <https://github.com/msiemens/tinydb/issues/7>`_: IDs not unique.
-- Extended the API: ``db.count(where(...))`` and ``db.contains(where(...))``
+- Extended the API: ``db.count(where(...))`` and ``db.contains(where(...))``.
 - The syntax ``query in db`` is now **deprecated** and replaced
   by ``db.contains``.